+++ .storybook/.babelrc
@@ -1,4 +1,4 @@
 {
-  "plugins": ["@compiled/babel-plugin-css-in-js"],
+  "plugins": ["@compiled/babel-plugin"],
   "presets": ["@babel/preset-react"]
 }
