+++ examples/css-prop-static-object.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '@compiled/css-in-js';
+import '@compiled/core';
 import { hover } from './mixins/mixins';
 
 const inlineMixinFunc = () => ({
