+++ examples/styled-dynamic-object.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { styled } from '@compiled/css-in-js';
+import { styled } from '@compiled/core';
 
 export default {
   title: 'styled component dynamic object',
