+++ package.json
@@ -28,8 +28,8 @@
   },
   "devDependencies": {
     "@babel/preset-react": "^7.10.1",
-    "@compiled/css-in-js": "*",
-    "@compiled/jest-css-in-js": "*",
+    "@compiled/core": "*",
+    "@compiled/jest": "*",
     "@compiled/ts-transform-css-in-js": "^0.4.7",
     "@size-limit/preset-small-lib": "^4.4.5",
     "@storybook/addons": "^5.2.8",
@@ -75,7 +75,7 @@
   },
   "size-limit": [
     {
-      "path": "./packages/style/dist/browser/index.js",
+      "path": "./packages/runtime/dist/browser/index.js",
       "limit": "299B",
       "import": "{ CC, CS }",
       "ignore": [
