+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@compiled/style",
+  "name": "@compiled/core",
   "version": "0.4.7",
   "description": "The CSS in JS authoring experience you love without the runtime cost",
   "author": "Michael Dougall",
@@ -9,18 +9,34 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/atlassian-labs/compiled-css-in-js.git",
-    "directory": "packages/style"
+    "directory": "packages/core"
   },
   "main": "./dist/index.js",
   "module": "./dist/index.js",
-  "browser": "./dist/browser/index.js",
   "types": "./dist/index.d.ts",
   "sideEffects": false,
+  "keywords": [
+    "compiled",
+    "css-in-js",
+    "styled-components",
+    "emotion-js",
+    "typescript",
+    "ts-transformer"
+  ],
   "files": [
     "dist",
     "src"
   ],
+  "dependencies": {
+    "@compiled/runtime": "0.4.7",
+    "csstype": "^2.0.0"
+  },
   "peerDependencies": {
     "react": "^16.12.0"
+  },
+  "devDependencies": {
+    "@testing-library/react": "^10.0.4",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
   }
 }
