+++ packages/core/src/index.tsx
@@ -1,4 +1,4 @@
-export { CS, CC } from '@compiled/style';
+export { CS, CC } from '@compiled/runtime';
 export { styled } from './styled';
 export { ClassNames } from './class-names';
 import { CssFunction } from './types';
