+++ packages/core/tsconfig.json
@@ -5,9 +5,9 @@
     "outDir": "dist",
     "baseUrl": "src",
     "paths": {
-      "@compiled/css-in-js": ["index.tsx"],
-      "@compiled/css-in-js/jsx": ["jsx.tsx"]
+      "@compiled/core": ["index.tsx"],
+      "@compiled/core/jsx": ["jsx.tsx"]
     }
   },
-  "references": [{ "path": "../style" }]
+  "references": [{ "path": "../runtime" }]
 }
