--- packages/css-in-js/README.md
@@ -1,1 +1,0 @@
-# @compiled/css-in-js
