+++ packages/runtime/README.md
@@ -1,3 +1,3 @@
-# @compiled/style
+# @compiled/runtime
 
 The _actual_ runtime for compiled.
