+++ packages/tsconfig.browser.json
@@ -1,4 +1,4 @@
 {
   "files": [],
-  "references": [{ "path": "style/tsconfig.browser.json" }]
+  "references": [{ "path": "runtime/tsconfig.browser.json" }]
 }
